+++ launcher/MainDlg.h
@@ -5,6 +5,8 @@
 #include <wxx_controls.h>
 #include <wxx_stdcontrols.h>
 #include <memory>
+#include <thread>
+#include <atomic>
 #include <launcher_common/UpdateChecker.h>
 #include "ImageButton.h"
 
@@ -30,6 +32,9 @@ private:
     void OnSupportLinkClick(int link_id);
     void OnOpenGameFolder(int folder_id);
     void OnAboutLinkClick();
+    void OnBnClickedFFLinkBtn();
+    LRESULT OnFFLinkComplete(WPARAM wparam, LPARAM lparam);
+    LRESULT OnFFLinkCancelled(WPARAM wparam, LPARAM lparam);
     void RefreshModSelector();
     CString GetSelectedMod();
     void AfterLaunch();
@@ -40,6 +45,13 @@ private:
     void ClearWhatsNewFlag();
     CEdit m_about_link;
     bool m_about_link_hover = false;
+
+    // FactionFiles Link
+    std::string m_fflink_token;
+    std::unique_ptr<std::thread> m_fflink_poll_thread;
+    std::atomic<bool> m_fflink_polling_active{false};
+    std::string m_fflink_result_username;
+    std::string m_fflink_result_token;
 
 protected:
     // Controls
@@ -59,6 +71,8 @@ protected:
     ImageButton m_sm2_button;
     ImageButton m_sm3_button;
     ImageButton m_sm4_button;
+
+    ImageButton m_fflink_button;
 
     CToolTip m_tool_tip;
 };